Moved the get_command_list function to module cmdseq. Removed some unused imports.
