+++ src/cmdseq.rs
@@ -1,4 +1,6 @@
-// Module cmdseq.
+// Module cmdseq
+use string_utils::{ collect_between_white, count_white_space };
+
 pub struct CmdSeq {
     times_before_next: usize,
     cmd: String,
@@ -16,4 +18,21 @@ impl CmdSeq {
     pub fn get_cmd(&self) -> &str {
         &self.cmd
     }
+}
+
+pub fn get_command_list(command: &str) -> (Vec<CmdSeq>, usize) {
+    let mut command_list: Vec<CmdSeq> = Vec::new();
+    // The number of operations we would execute before we need to start from beginning.
+    let mut num_operations: usize = 0;
+    let mut num = 0;
+    while num <= count_white_space(command) { // Using 'while' loop as 'step_by()' has issues atm.
+        let t_b_n: usize = collect_between_white(command, num, num + 1).parse().expect("Usage: cmdseq [-d <count dir>] <count1> <cmd1> [... <countn> <cmdn>]");
+        num_operations += t_b_n;
+        command_list.push(CmdSeq::new(
+            t_b_n,
+            collect_between_white(command, num + 1, num + 2).replace("\"", ""),
+        ));
+        num = num + 2;
+    }
+    (command_list, num_operations)
 }
